+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Deprecated
 
 ### Removed
+- GetFeatureInfo: The attributes `imageLinks` and `maxWidth` have been removed from the gfiTheme `Default`, as the images are now rendered within the table.
 
 ### Fixed
 - Issue #1121: For secured services that have the isSecured attribute, the lock icon is now displayed again.